% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_Qtypes.R
\name{is_selectmultiple}
\alias{is_selectmultiple}
\title{Check that question is select_multiple as definied by XLSform}
\usage{
is_selectmultiple(df, column)
}
\arguments{
\item{df}{The dataframe object to be processed}

\item{column}{Column/question name to be checked, as string}
}
\value{
TRUE is question validates, FALSE otherwise
}
\description{
Check that question is select_multiple as definied by XLSform
}
