% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_Qtypes.R
\name{is_selectone}
\alias{is_selectone}
\title{Check that question is is_selectone as definied by XLSform}
\usage{
is_selectone(df, column, num_fact = 100)
}
\arguments{
\item{df}{The dataframe object to be processed}

\item{column}{Column/question name to be checked, as string}
}
\value{
TRUE is question validates, FALSE otherwise
}
\description{
Check that question is is_selectone as definied by XLSform
}
\author{
Elliott Messeiller
#'
}
