% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_api.R
\name{kobo_all_forms}
\alias{kobo_all_forms}
\title{Returns a dataframe with all forms available for the user}
\usage{
kobo_all_forms(user, pwd, api = "https://kobo.humanitarianresponse.info")
}
\arguments{
\item{user}{Optional. A single string indicating the username}

\item{pwd}{Optional. A single string indicating the password}

\item{api}{The URL at which the API can be accessed. Default to "kobo.humanitarianresponse.info"}
}
\value{
Dataframe with all forms available for the user
}
\description{
Download form from the platform
}
\author{
Elliott Messeiller
}
