% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_crunch.R
\name{kobo_crunching}
\alias{kobo_crunching}
\title{Pick and run analysis function for question type.}
\usage{
kobo_crunching(question, data, dicofile = "dico.csv")
}
\arguments{
\item{question}{Question to be treated by the function. It should be the "name" of the  questionas in the dictionnary or XLSform.}

\item{data}{Dataframe containing data to analyse}

\item{dicofile}{Name of the file containing the dictionnary. "dico.csv" is used by default.}
}
\value{
Returns a list containing four to five named elements:
\itemize{
\item variable : The name of the question (variable) analysed
\item plot: A ggplot2 object with a basic bar or density graph depending on the question
\item freq_tbl: A frequency table of the question of the question
\item summary: summary statistics (quartiles and mean). Only for decimal/integer questions.
\item other: Other information on the question: # respondents answered the form, # and % of respondents to the question.
}
}
\description{
This function pick and run the right analysis function depending on the type of question for data crunching.
}
