% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_crunch.R
\name{kobo_crunching_all}
\alias{kobo_crunching_all}
\title{Crunch all questions in the dictionnary}
\usage{
kobo_crunching_all(data, dicofile = "dico.csv")
}
\arguments{
\item{data}{Dataframe containing data to analyse}

\item{dicofile}{Name of the file containing the dictionnary. "dico.csv" is used by default.}
}
\value{
Returns a list containing four to five named elements:
\itemize{
\item variable : The name of the question (variable) analysed
\item plot: A ggplot2 object with a basic bar or density graph depending on the question
\item freq_tbl: A frequency table of the question of the question
\item summary: summary statistics (quartiles and mean). Only for decimal/integer questions.
\item other: Other information on the question: # respondents answered the form, # and % of respondents to the question.
}
}
\description{
Run kobo_crunching on all the questions present in the dictionnary and the dataframe.
}
