% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_api.R
\name{kobo_data}
\alias{kobo_data}
\title{Download data from the platform}
\usage{
kobo_data(formid, user, pwd,
  api = "https://kobo.humanitarianresponse.info")
}
\arguments{
\item{formid}{The ID of the form to be accessed (as a character string). It must in the API V1 format: see kobo_all_forms old_id output.}

\item{user}{Optional. A single string indicating the username}

\item{pwd}{Optional. A single string indicating the password}

\item{api}{The URL at which the API can be accessed. Default to "kobo.humanitarianresponse.info"}
}
\value{
The file path where the csv file was saved.
}
\description{
Download data from the platform
}
\author{
Elliott Messeiller
}
