% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_dico.R
\name{kobo_dico}
\alias{kobo_dico}
\title{Data dictionnary}
\usage{
kobo_dico(form, dicofile = "dico.csv")
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the working folder.}

\item{dicofile}{Name of the csv file to be created. It must end with ".csv". Path to another folder in the working directory can be inserted.}
}
\value{
Write a csv file in the working directory with the name given in the dicofile argument. Path to another folder in the working directory can be inserted in the dicofile argument.
}
\description{
Produce a data dictionnary based on the xlsform for the project
}
\author{
Edouard Legoupil, Elliott Messeiller
}
