% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_crunch.R
\name{kobo_histo_indiv}
\alias{kobo_histo_indiv}
\title{Generate histograme for individual integer questions}
\usage{
kobo_histo_indiv(data, question, dicofile = "dico.csv")
}
\arguments{
\item{data}{Dataframe containing data to analyse}

\item{question}{Question to be treated by the function. It should be the "name" of the questions as in the dictionnary or XLSform.
The question has to be a decimal or integer question in the dictionary.}

\item{dicofile}{Name of the file containing the dictionnary. "dico.csv" is used by default.}
}
\value{
Returns a list containing four elements:
\itemize{
\item variable : The name of the question (variable) analysed
\item plot: A ggplot2 object with a basic bar or density graph depending on the question
\item freq_tbl: A frequency table of the question of the question
\item summary: summary statistics (quartiles and mean). Only for decimal/integer questions.
\item other: Other information on the question: # respondents answered the form, # and % of respondents to the question.
}
}
\description{
Automatically generate histogrammes for each of the integer questions in the dataset. ggplot2 is used.
}
\examples{
kobo_histo_indiv()
\dontrun{
kobo_histo_indiv()
}

}
\author{
Edouard Legoupil, Elliott Messeiller
}
