% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_label.R
\name{kobo_label}
\alias{kobo_label}
\title{Label Variable}
\usage{
kobo_label(data, dico)
}
\arguments{
\item{data}{Data where labels must be inserted}

\item{dico}{( generated from kobo_dico)}
}
\value{
A "data.table" with the full data.label, and choices labels. To be used for graphs generation.
}
\description{
Insert the full label in data frame based on dictionnary
}
\author{
Edouard Legoupil
}
