% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_xlsform_relabel.R
\name{kobo_xlsform_relabel}
\alias{kobo_xlsform_relabel}
\title{Relabel XLSform based on an original dataframe}
\usage{
kobo_xlsform_relabel(new_form, original_form, name_form = "form.xls")
}
\arguments{
\item{new_form}{The full filename of the form to be relabeled (xls or xlsx file).}

\item{original_form}{The full filename of the form to match the assessed form.}
}
\description{
Match labels from an XLSform created with the kobo_to_xlsform function with a prexisting XLSform.
It can then be used to match and relabel cleaned datasets with original forms.
}
\examples{
data(iris)
str(iris)
kobo_to_xlsform(iris)

}
\author{
Elliott Messeiller
}
